+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2.0.0 (27/05/2020)
+
+### Breaking changes
+ * Updated dependecy to `@animoca/ethereum-contracts-core_library:2.0.0`.
+ * Updated compiler to `solc:0.6.8`.
+ * Added `ERC20.sol` and `ERC20WithOperators.sol`. Removed `ERC20Full.sol`. Refactored the unit tests.
+ * Applied naming convention on public members.
+
+### Improvements
+ * Support of new ERC165 interface for allowance functions.
+
 ## 1.0.1 (06/05/2020)
 
 ### Improvements